edgedetector/requirements.txt
# https://github.com/dpkp/kafka-python
kafka-python

# https://github.com/jmcarp/py-cfenv
cfenv

opencv-python
numpy
imutils
